Splash Screen merged with bottom navbar
